1. 清除warning 2. 补充之前新加的timestamp文件
git-svn-id: http://192.168.20.11/sequoiadb/trunk@21083 6121f079-4a18-41d3-8c30-8870a322f2c2

+++ SequoiaDB/engine/client/jstobs.c
@@ -528,12 +528,12 @@ static BOOLEAN bsonConvertJson ( CHAR **pbuf,
             _snprintf ( temp,
                         BSON_TEMP_SIZE_512,
                         "%lld",
-                        timer * 1000 ) ;
+                        (UINT64)(timer * 1000) ) ;
 #else
             snprintf ( temp,
                        BSON_TEMP_SIZE_512,
                        "%lld",
-                       timer * 1000 ) ;
+                       (UINT64)(timer * 1000) ) ;
 #endif
             bsonConvertJsonRawConcat ( pbuf, left, "{ \"$date\": ", FALSE ) ;
             CHECK_LEFT ( left )
